Fix missing type annotation
